+++ packages/create-sitecore-jss/src/templates/node-xmcloud-proxy/src/index.ts
@@ -1,11 +1,11 @@
 import 'dotenv/config';
 import express, { Response } from 'express';
 import compression from 'compression';
-import { createProxyMiddleware } from 'http-proxy-middleware';
+import { createProxyMiddleware, fixRequestBody } from 'http-proxy-middleware';
 import { debug } from '@sitecore-jss/sitecore-jss';
-import { editingRouter } from '@sitecore-jss/sitecore-jss-proxy';
-import { healthCheck } from '@sitecore-jss/sitecore-jss-proxy';
-import { config } from './config';
+import { editingRouter, healthCheck } from '@sitecore-jss/sitecore-jss-proxy';
+import { config, graphQLEndpoint } from './config';
+import { personalizeHelper, personalizePlugin } from './personalize';
 
 const server = express();
 
@@ -45,31 +45,6 @@ const layoutService = layoutServiceFactory.create();
 
 const dictionaryService = dictionaryServiceFactory.create();
 
-const clientFactoryConfig = config.serverBundle.getClientFactoryConfig();
-
-/**
- * GraphQL endpoint resolution to meet the requirements of the http-proxy-middleware
- */
-const graphQLEndpoint = (() => {
-  try {
-    const graphQLEndpoint = new URL(clientFactoryConfig.endpoint);
-    // Browser request path to the proxy. Includes only the pathname.
-    const pathname = graphQLEndpoint.pathname;
-    // Target URL for the proxy. Can't include the query string.
-    const target = `${graphQLEndpoint.protocol}//${graphQLEndpoint.hostname}${pathname}`;
-
-    return {
-      target,
-      path: pathname,
-    };
-  } catch (error) {
-    throw new Error(
-      `ERROR: The serverBundle should export a getClientFactoryConfig function with valid GraphQL endpoint URL returned, current value is ${clientFactoryConfig.endpoint}. ` +
-        'Please check your server bundle.'
-    );
-  }
-})();
-
 /**
  * Parse requested url in order to detect current route and language
  * @param {string} reqRoute requested route
@@ -105,6 +80,8 @@ const handleError = (res: Response, err: unknown) => {
 
 // enable gzip compression for appropriate file types
 server.use(compression());
+// enable access to req.body
+server.use(graphQLEndpoint.path, express.json());
 
 // turn off x-powered-by http header
 server.settings['x-powered-by'] = false;
@@ -125,6 +102,12 @@ server.use(
   createProxyMiddleware({
     target: graphQLEndpoint.target,
     changeOrigin: true,
+    selfHandleResponse: true,
+    on: {
+      proxyReq: fixRequestBody,
+    },
+    // for client-side routing, personalization is performed by modifying layout service response
+    plugins: [personalizePlugin],
   })
 );
 
@@ -165,11 +148,17 @@ server.use(async (req, res) => {
     }
 
     // Language is required. In case it's not specified in the requested URL, fallback to the default language from the app configuration.
-    const layoutData = await layoutService.fetchLayoutData(
+    let layoutData = await layoutService.fetchLayoutData(
       route,
       lang || config.serverBundle.defaultLanguage
     );
-
+    // for SSR loading routing, personalization is performed by modifying layoutData directly
+    const personalizedLayoutData = await personalizeHelper.personalizeLayoutData(
+      req,
+      res,
+      layoutData
+    );
+    layoutData = personalizedLayoutData;
     const viewBag = { dictionary: {} };
 
     viewBag.dictionary = await dictionaryService.fetchDictionaryData(