+++ packages/sitecore-jss-proxy/src/index.ts
@@ -1,2 +1,4 @@
 export * from './middleware';
 export * from './types';
+export * from './personalize';
+export { GRAPHQL_LAYOUT_QUERY_NAME } from '@sitecore-jss/sitecore-jss/layout';